fields/timestamp: fix Timestmap.Truncate issues
Fix issues with the Timestmap.Truncate method, and add relevant tests,
including Fuzz tests. The main issue was that it was missing a division
by 1e3 to get the value in the same resolution as the Timestamp type.

Secondly, there where semantical issues at and around the origin time,
as well as at and around the UNIX zero time. The cases where found via
Fuzzing, and added to the normal tests afterwards.