automationcli: improve help text (breaking)
Improve the help text for the automationcli pacakge, including print of
flags and tree structure of available routines.

This require a breaking change to the automationcli.Config type, as it's
required to include availalbe routines in order to be able to print
them. Programs that use this struct directly, or call ParseArguments
directly, would need to be updated.

Programs that call the ParseAndRun helper function are unafected.